ourIndia_backend/requirements.txt
anyio==3.7.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.7.2
asttokens==2.2.1
async-lru==2.0.4
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
cachetools==5.3.3
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.3
colorama==0.4.6
comm==0.1.4
cryptography==42.0.5
debugpy==1.6.8
decorator==5.1.1
defusedxml==0.7.1
Django==5.0.3
django-cors-headers==4.3.1
django-dotenv==1.4.2
django-filter==24.1
django-templated-mail==1.1.1
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
docx2txt==0.8
drf-spectacular==0.27.2
et-xmlfile==1.1.0
executing==1.2.0
fastjsonschema==2.18.0
Flask==2.2.2
fqdn==1.5.1
google-api-core==2.18.0
google-api-python-client==2.124.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.63.0
httplib2==0.22.0
idna==3.4
inflection==0.5.1
ipykernel==6.25.0
ipython==8.14.0
ipython-genutils==0.2.0
ipywidgets==8.1.0
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.18.6
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.7.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.3
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.8
jupyterlab_server==2.24.0
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mistune==3.0.1
mysqlclient==2.2.4
nbclient==0.8.0
nbconvert==7.7.3
nbformat==5.9.2
nest-asyncio==1.5.7
notebook==7.0.1
notebook_shim==0.2.3
numpy==1.25.2
oauthlib==3.2.2
openpyxl==3.1.2
overrides==7.3.1
packaging==23.1
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.10.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.5
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.21
Pygments==2.15.1
PyJWT==2.8.0
pyparsing==3.1.2
PyPDF2==3.0.1
python-dateutil==2.8.2
python-json-logger==2.0.7
python3-openid==3.2.0
pytz==2023.3
pywin32==306
pywinpty==2.0.11
PyYAML==6.0.1
pyzmq==25.1.0
qtconsole==5.4.3
QtPy==2.3.1
referencing==0.30.0
requests==2.31.0
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.9.2
rsa==4.9
scikit-learn==1.3.0
scipy==1.11.2
Send2Trash==1.8.2
six==1.16.0
sniffio==1.3.0
social-auth-app-django==5.4.0
social-auth-core==4.5.3
soupsieve==2.4.1
sqlparse==0.4.3
stack-data==0.6.2
terminado==0.17.1
threadpoolctl==3.2.0
tinycss2==1.2.1
tornado==6.3.2
traitlets==5.9.0
tzdata==2022.7
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.0.4
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.1
Werkzeug==2.2.2
widgetsnbextension==4.0.8

# pip freeze > requirements.txt
